factnet/backend/requirements.txt
# FactNet Backend Dependencies

# Django
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7

# MongoDB
djongo==1.3.6
pymongo==4.6.0
dnspython==2.4.2

# Social Auth
social-auth-app-django==5.4.0
social-auth-core==4.5.0

# NLP (for similarity checking)
sentence-transformers==2.3.0
torch>=2.0.0
nltk==3.8.1

# Utilities
python-dotenv==1.0.0